# app/src/lib/hub/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "illyHub hub",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/transport/{action}": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Transport",
    "operationId": "transport_api_transport__action__post",
    "parameters": [
     {
      "name": "action",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "play",
        "pause",
        "toggle",
        "stop",
        "next",
        "prev"
       ],
       "type": "string",
       "title": "Action"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TargetBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/seek": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Seek",
    "operationId": "seek_api_seek_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SeekBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/skip": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Skip",
    "operationId": "skip_api_skip_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SkipBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/volume": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Volume",
    "operationId": "volume_api_volume_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VolumeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/mute": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Mute",
    "operationId": "mute_api_mute_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MuteBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/zone/power": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Zone Power",
    "operationId": "zone_power_api_zone_power_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ZonePowerBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/group": {
   "post": {
    "tags": [
     "commands"
    ],
    "summary": "Group",
    "operationId": "group_api_group_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GroupBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/group/{side_id}": {
   "delete": {
    "tags": [
     "commands"
    ],
    "summary": "Ungroup",
    "operationId": "ungroup_api_group__side_id__delete",
    "parameters": [
     {
      "name": "side_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Side Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ack"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/dev/fake/{scenario}": {
   "post": {
    "tags": [
     "dev"
    ],
    "summary": "Fake Scenario",
    "operationId": "fake_scenario_api_dev_fake__scenario__post",
    "parameters": [
     {
      "name": "scenario",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Scenario"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScenarioResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "Ack": {
    "properties": {
     "correlation_id": {
      "type": "string",
      "title": "Correlation Id"
     },
     "ok": {
      "type": "boolean",
      "title": "Ok"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "target": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target"
     },
     "state_version": {
      "type": "integer",
      "title": "State Version"
     },
     "error": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ErrorEnvelope"
       },
       {
        "type": "null"
       }
      ]
     },
     "partial": {
      "items": {
       "$ref": "#/components/schemas/PartialFailure"
      },
      "type": "array",
      "title": "Partial"
     }
    },
    "type": "object",
    "required": [
     "correlation_id",
     "ok",
     "action",
     "state_version"
    ],
    "title": "Ack"
   },
...
   "ErrorEnvelope": {
    "properties": {
     "code": {
      "type": "string",
      "enum": [
       "unknown_target",
       "unsupported_action",
       "not_seekable",
       "adapter_disconnected",
       "device_offline",
       "invalid_argument",
       "vendor_error"
      ],
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "target": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target"
     },
     "correlation_id": {
      "type": "string",
      "title": "Correlation Id"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "correlation_id"
    ],
    "title": "ErrorEnvelope"
   },
   "GroupBody": {
    "properties": {
     "vendor": {
      "type": "string",
      "enum": [
       "heos",
       "sonos"
      ],
      "title": "Vendor"
     },
     "coordinator_id": {
      "type": "string",
      "title": "Coordinator Id"
     },
     "member_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Member Ids"
     }
    },
    "type": "object",
    "required": [
     "vendor",
     "coordinator_id",
     "member_ids"
    ],
    "title": "GroupBody"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MuteBody": {
    "properties": {
     "target": {
      "type": "string",
      "title": "Target",
      "description": "player id, side id, or 'all'"
     },
     "muted": {
      "type": "boolean",
      "title": "Muted"
     }
    },
    "type": "object",
    "required": [
     "target",
     "muted"
    ],
    "title": "MuteBody"
   },
   "PartialFailure": {
    "properties": {
     "target": {
      "type": "string",
      "title": "Target"
     },
     "code": {
      "type": "string",
      "enum": [
       "unknown_target",
       "unsupported_action",
       "not_seekable",
       "adapter_disconnected",
       "device_offline",
       "invalid_argument",
       "vendor_error"
      ],
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "target",
     "code",
     "message"
    ],
    "title": "PartialFailure"
   },
...
   "ScenarioResponse": {
    "properties": {
     "scenario": {
      "type": "string",
      "title": "Scenario"
     },
     "result": {
      "additionalProperties": true,
      "type": "object",
      "title": "Result"
     },
     "state_version": {
      "type": "integer",
      "title": "State Version"
     }
    },
    "type": "object",
    "required": [
     "scenario",
     "result",
     "state_version"
    ],
    "title": "ScenarioResponse"
   },
   "SeekBody": {
    "properties": {
     "target": {
      "type": "string",
      "title": "Target",
      "description": "player id, side id, or 'all'"
     },
     "position_ms": {
      "type": "integer",
      "minimum": 0.0,
      "title": "Position Ms"
     }
    },
    "type": "object",
    "required": [
     "target",
     "position_ms"
    ],
    "title": "SeekBody"
   },
...
   "SkipBody": {
    "properties": {
     "target": {
      "type": "string",
      "title": "Target",
      "description": "player id, side id, or 'all'"
     },
     "delta_ms": {
      "type": "integer",
      "title": "Delta Ms",
      "description": "signed; \u00b115000 for the transport row",
      "default": 15000
     }
    },
    "type": "object",
    "required": [
     "target"
    ],
    "title": "SkipBody"
   },
   "TargetBody": {
    "properties": {
     "target": {
      "type": "string",
      "title": "Target",
      "description": "player id, side id, or 'all'"
     }
    },
    "type": "object",
    "required": [
     "target"
    ],
    "title": "TargetBody"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VolumeBody": {
    "properties": {
     "target": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target"
     },
     "level": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Level"
     },
     "linked": {
      "type": "boolean",
      "title": "Linked",
      "default": false
     },
     "delta": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Delta"
     }
    },
    "type": "object",
    "title": "VolumeBody"
   },
...
   "ZonePowerBody": {
    "properties": {
     "zone_id": {
      "type": "string",
      "title": "Zone Id",
      "description": "Denon zone id, or a Sonos player/side id for stop+ungroup"
     },
     "on": {
      "type": "boolean",
      "title": "On"
     }
    },
    "type": "object",
    "required": [
     "zone_id",
     "on"
    ],
    "title": "ZonePowerBody"
   }
  }
 },
...
}